NIMAS-specific schema adjustments
Some educated guesses, based on discussions with APH. I am waiting for
confirmation from them but I think this will be in line with their
needs. I'm submitting it now because I'm going to be out of office for
the next 2 weeks and I know we have a release coming up.