initial commit for a XProc/XSLT port of the Pipeline 1 DAISY 2.02 validator
More work on the EPUB3 to DAISY 2.02 converter made it clear that a DAISY 2.02 validator is very useful.
A DAISY 2.02 validator will be invoked after conversion to make sure the result is valid. The EPUB3 Validator (epubcheck)
will also be invoked before conversion to make sure the input is valid.

The port of the "Daisy 2.02 DTB Light Validator" is rudamentary and untested but it should run successfully.
The only thing it validates at the moment is that the SMIL clock values add up to the declared totalTime in the SMIL and NCC headers.