css:inline: add 'context' input port for URI resolving
Also fixed a bug in getting the base URI of the input document. The use
case is for unzipped files which have empty base URI, but for which an
xml:base attribute has been added to the document element.

To do: check whether the base URI fix is still needed with
https://github.com/daisy/pipeline-modules-common/pull/101.