Fix MathML ID conversion in dtbook-to-zedai
ZedAI schemadoc seems to say `@id` is not allowed on MathML elements
so we discarded it (and it resulted in some IDs lost during the
conversion).
It appears that the MathML feature's schemas do allow `@id` (in place
of ZedAI's usual `@xml:id`), so we now keep those during the conversion,
which consequently fixes ID-loss issues.

This change also gets rid of `@dtbook:*` attributes in MathML markup.