A Transform.Provider can now be given a "job context"
The main purpose is to be able to create job messages from Java.

JobContext currently implements the SLF4J Logger interface. Messages
sent to this logger should be forwarded to the framework's message event
bus.

The idea is that eventually we'll get JobContext objects from the
framework, but for the time being I construct my own JobContext that
uses a hack with Calabash's XProcMessageListener.

See issue https://github.com/daisy/pipeline-mod-braille/issues/19