Don't ignore soft hyphens in dotify:format if hyphenate=false
because the attribute is not being set to true css-to-obfl.

Also, according to the OBFL specification hyphenate="true" corresponds
to "hyphens:auto", while hyphenate="false" (default) corresponds to
"hyphens:manual". This means that in order to support hyphens:auto, I'll
need to add hyphenate attributes in css-to-obfl, and in order to support
hyphens:none, I'll need to remove soft hyphens in css-to-obfl.

see issue https://github.com/daisy/pipeline-mod-braille/issues/33