Remove some trailing blank patterns and trailing empty rows in expected PEFs
because org.daisy.braille.pef.PEFFileCompare chokes on it.

In the cases where Dotify generates trailing blank patterns and trailing
empty rows, I have intentionally not fixed the expected PEFs because it
would make the tests less readable (and it wouldn't be a real fix
because the tests would still be brittle). So instead we should enhance
PEFFileCompare.

see https://github.com/joeha480/dotify/issues/114