Make it possible to derive a script from a main script
This is done by using the "px:extends-script" attribute in
catalog.xml. It causes all attributes and documentation defined on
inputs and options of the main scripts to be automatically copied to the
derived script. Any attributes or documentation defined on the derived
script are not overriden. This mechanism can reduce a script to a simple
configuration file for another script for setting default option values,
hiding options, changing allowed option values, etc.