generate-toc.xsl: rename _depth parameter to depth
and rename the internal variable to _depth. The internal variable is
needed because XProc can only pass parameters as strings so if we need
other types such as boolean we need to create additional variables. We
can have variable names that start with an underscore but it's better to
use these for the internal variables and not the parameter.