From now on, don't comment out modules in aggregator POMs
which will simplify the workflow a lot.

The reason for disabling modules was two-fold:
- to be able to release only a part of the modules (with a single
  command), and
- build only what is strictly needed to avoid long build times

We can accomplish the same with the maven switch `--projects <list of
modules>`. The list can easily be generated from the BOM by selecting
the modules with SNAPSHOT versions.

One could go one step further and only select modules with SNAPSHOT
versions and that have been changed *since the last build* (e.g. by
comparing the modification times of the source files vs. JAR files under
~/.m2).

Another benefit of this approach is that modules that haven't changed
but depend on changed modules can be automatically built and tested
against the newest versions, by using the `--also-make-dependents`
switch.