Remove "no block allowed within block with underline properties" restriction
According to OBFL the behavior is unspecified, but it is OK to allow
it. If a block with underline properties contains an empty block with
text following but not preceding, or preceding but not following, the
behavior is as if the empty block was not present.

see https://github.com/brailleapps/dotify.formatter.impl/pull/17