Update release procedure in README
Because we don't permanently comment out modules in aggregator POMs
anymore, we have to do it temporarily every time we make a release. For
now this needs to be done manually.

Using the --projects option to specify the modules that we want to
release doesn't work for some reason. (Advanced reactor options like
"--projects" don't seem to be compatible with aggregator plugins like
the maven-release-plugin?)