New version of XSLT implementation of css:split
The `preceding::` and `following::` axes seem to have caused performance
issues. This new implementation uses recursion and is significantly
faster. A special recursion method is used in order to not cause stack
overflows. Note that with XSLT 3.0's xsl:iterate the code could be
simplified a lot.

Also copied the xprocspec test to xspec.