Update to latest modules-parent
How registering of modules works has changed. Instead of the framework
tracking bundles and processing modules, the modules are now
responsible for exposing theirselves as Module services.

Also make the _dsannotations instructions explicit everywhere. This
makes it immediately clear which services a bundle exposes, including
the module and script services.

_dsannotations is specified through a property "expose-services", so
that it can be reused later for the ds-to-spi-maven-plugin.

Also fix "split-package" warnings by adding Private-Package
instructions with a "split-package" directive.