daisy202-validator: fix #82: attribute "shape" not allowed here
See issue https://github.com/daisy/pipeline-scripts/issues/82

The shape attribute is automatically added because of the DTD
default (see
https://github.com/daisy/pipeline-scripts-utils/blob/f297dec/common-entities/src/main/resources/xml/xhtml1-transitional.dtd#L619)
after which the RNG for validating the NCC (d202ncc.rng) complaints.

I fixed it by simply uncommenting a line in d202ncc.rng. I'm not sure
why it was commented out. The code was ported from Pipeline 1 (see
https://github.com/daisy/pipeline-scripts/commit/2b04ed2) so I'd have
to look it up in the Pipeline 1 history.