Rename px:x-to-y-convert to px:x-to-y and px:x-to-y to px:x-to-y.script
The reasoning is that the "convert" step is the main step from the
point of view of the XProc library user, and that the script step is a
thin Pipeline specific wrapper that doesn't even need to be in the
XProc library.

Also added convert scripts wherever they were not present yet.